Extract converting product list to new class
