+++ app/src/main/java/com/ratajczykdev/inventoryapp/dataexport/PdfExportFragment.kt
@@ -25,10 +25,6 @@ import java.util.*
 class PdfExportFragment : Fragment() {
     //  change to export PDF not TXT
 
-    private companion object {
-        const val STRING_FORMAT_STYLE = "%-10s %-25s %-15s %-10s"
-    }
-
     private lateinit var productViewModel: ProductViewModel
     private var productList: MutableList<Product> = mutableListOf()
 
@@ -48,28 +44,10 @@ class PdfExportFragment : Fragment() {
     private fun action() {
         //  TODO: change method name
         val externalExportDir = StorageOperations.createDirInExternal("exported", context)
-        val stringProductList = createStringProductList()
+        val stringProductList = ProductListExporter.createStringProductList(productList)
         StorageOperations.writeStringToFile(externalExportDir, "exported_database.txt", stringProductList)
     }
 
-    private fun createStringProductList(): String {
-        var stringProductList = createDatabaseHeader() + "\n"
-        for (product in productList) {
-            val id = product.id.toString()
-            val name = product.name.toString()
-            val price = product.price.toString()
-            val quantity = product.quantity.toString()
 
-            stringProductList += String.format(STRING_FORMAT_STYLE, id, name, price, quantity) + "\n"
-        }
-        return stringProductList
-    }
-
-    private fun createDatabaseHeader(): String {
-        val dateLine = Calendar.getInstance().time.toString() + "\n"
-        val columnNames = String.format(STRING_FORMAT_STYLE,
-                "[ID]", "[NAME]", "[PRICE]", "[QUANTITY]")
-        return dateLine + columnNames
-    }
 
 }
